Rename Streamly.Internal.Memory.Array.Types
According to the new module structure, it should be Streamly.Internal.Data.Array.Storable.Foreign.Types
