+++ benchmark/Streamly/Benchmark/FileSystem/Handle/ReadWrite.hs
@@ -49,7 +49,7 @@ import Streamly.Internal.Data.Stream.StreamD.Type (Step(..))
 import qualified Streamly.Internal.Data.Stream.StreamD.Type as D
 import qualified Streamly.Internal.Data.Tuple.Strict as Strict
 import qualified Streamly.Internal.Memory.Array as A
-import qualified Streamly.Internal.Memory.Array.Types as AT
+import qualified Streamly.Internal.Data.Array.Storable.Foreign.Types as AT
 
 import Test.Inspection
 #endif
