+++ src/Streamly/Internal/FileSystem/Dir.hs
@@ -63,7 +63,7 @@ import Prelude hiding (read)
 
 -- import Streamly.Data.Fold (Fold)
 import Streamly.Internal.Data.Unfold.Types (Unfold(..))
--- import Streamly.Internal.Memory.Array.Types
+-- import Streamly.Internal.Data.Array.Storable.Foreign.Types
 --        (Array(..), writeNUnsafe, defaultChunkSize, shrinkToFit,
 --         lpackArraysChunksOf)
 -- import Streamly.Internal.Data.Stream.Serial (SerialT)
@@ -177,9 +177,9 @@ toChunks = toChunksWithBufferOf defaultChunkSize
 
 -- | Unfolds a handle into a stream of 'Word8' arrays. Requests to the IO
 -- device are performed using a buffer of size
--- 'Streamly.Internal.Memory.Array.Types.defaultChunkSize'. The
+-- 'Streamly.Internal.Data.Array.Storable.Foreign.Types.defaultChunkSize'. The
 -- size of arrays in the resulting stream are therefore less than or equal to
--- 'Streamly.Internal.Memory.Array.Types.defaultChunkSize'.
+-- 'Streamly.Internal.Data.Array.Storable.Foreign.Types.defaultChunkSize'.
 --
 -- @since 0.7.0
 {-# INLINE readChunks #-}
@@ -354,7 +354,7 @@ fromStreamWithBufferOf n h m = fromChunks h $ S.arraysOf n m
 -- > write = 'writeWithBufferOf' A.defaultChunkSize
 --
 -- | Write a byte stream to a file handle. Accumulates the input in chunks of
--- up to 'Streamly.Internal.Memory.Array.Types.defaultChunkSize' before writing.
+-- up to 'Streamly.Internal.Data.Array.Storable.Foreign.Types.defaultChunkSize' before writing.
 --
 -- NOTE: This may perform better than the 'write' fold, you can try this if you
 -- need some extra perf boost.
@@ -404,7 +404,7 @@ writeWithBufferOf n h = FL.lchunksOf n (writeNUnsafe n) (writeChunks h)
 -- > write = 'writeWithBufferOf' A.defaultChunkSize
 --
 -- | Write a byte stream to a file handle. Accumulates the input in chunks of
--- up to 'Streamly.Internal.Memory.Array.Types.defaultChunkSize' before writing
+-- up to 'Streamly.Internal.Data.Array.Storable.Foreign.Types.defaultChunkSize' before writing
 -- to the IO device.
 --
 -- @since 0.7.0
