+++ src/Streamly/Internal/FileSystem/Event/Linux.hs
@@ -161,7 +161,7 @@ import GHC.IO.FD (fdFD, mkFD)
 import GHC.IO.Handle.FD (mkHandleFromFD)
 import Streamly (SerialT)
 import Streamly.Internal.Data.Parser (Parser)
-import Streamly.Internal.Memory.Array.Types (Array(..))
+import Streamly.Internal.Data.Array.Storable.Foreign.Types (Array(..))
 import System.IO (Handle, hClose, IOMode(ReadMode))
 #if !MIN_VERSION_base(4,10,0)
 import Control.Concurrent.MVar (readMVar)
