+++ src/Streamly/Internal/Network/Inet/TCP.hs
@@ -109,7 +109,7 @@ import Streamly.Internal.Data.SVar (fork)
 import Streamly.Internal.Data.Unfold.Types (Unfold(..))
 import Streamly.Internal.Network.Socket (SockSpec(..), accept, connections)
 import Streamly.Internal.Data.Stream.Serial (SerialT)
-import Streamly.Internal.Memory.Array.Types (Array(..), defaultChunkSize, writeNUnsafe)
+import Streamly.Internal.Data.Array.Storable.Foreign.Types (Array(..), defaultChunkSize, writeNUnsafe)
 import Streamly.Internal.Data.Stream.StreamK.Type (IsStream)
 
 import qualified Control.Monad.Catch as MC
