+++ src/Streamly/Network/Socket.hs
@@ -64,9 +64,9 @@
 -- = Programmer Notes
 --
 -- Read IO requests to connected stream sockets are performed in chunks of
--- 'Streamly.Internal.Memory.Array.Types.defaultChunkSize'.  Unless specified
+-- 'Streamly.Internal.Data.Array.Storable.Foreign.Types.defaultChunkSize'.  Unless specified
 -- otherwise in the API, writes are collected into chunks of
--- 'Streamly.Internal.Memory.Array.Types.defaultChunkSize' before they are
+-- 'Streamly.Internal.Data.Array.Storable.Foreign.Types.defaultChunkSize' before they are
 -- written to the socket. APIs are provided to control the chunking behavior.
 --
 -- > import qualified Streamly.Network.Socket as SK
