+++ streamly.cabal
@@ -430,7 +430,7 @@ library
                      , Streamly.Internal.Data.SmallArray.Types
                      , Streamly.Internal.Data.SmallArray
                      , Streamly.Internal.Data.Array.Storable.Foreign.Mut.Types
-                     , Streamly.Internal.Memory.Array.Types
+                     , Streamly.Internal.Data.Array.Storable.Foreign.Types
                      , Streamly.Internal.Memory.Array
                      , Streamly.Internal.Memory.ArrayStream
